requirements.txt
PyQt5
GitPython
requests
